nelogoutina admino jei kaa

+++ app/Http/Controllers/CartController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Kategorija;
+use App\Krepselis;
 use App\Nuotrauka;
 use App\Preke;
 use App\PrekeKrepselis;
@@ -15,17 +16,15 @@ class CartController extends Controller
 {
     public function index(){
         $allcategories=Kategorija::all();
-        $item = Preke::all();
         $kr=session('krepselis');
-//        dd($kr);
         if(session()->has('krepselis')) {
             $result=DB::table('krepselis')->where('krepselis.id_krepselis','=',$kr)->leftJoin('preke_krepselis', 'id_krepselis','=','preke_krepselis.fk_krepselis')
                 ->leftJoin('preke','preke_krepselis.fk_preke','=','id_preke')
                 ->select('preke_krepselis.*','preke.kaina','preke.pavadinimas','preke.aprasymas',DB::raw('krepselis.kaina as kr_kaina'))->get();
 
-            $kiekis = DB::table('preke_krepselis')->where('preke_krepselis.fk_krepselis','=',$kr)->get();
-           $kiekelis=0;
-            foreach ($kiekis as $kk){
+            $visosp = DB::table('preke_krepselis')->where('preke_krepselis.fk_krepselis','=',$kr)->get();
+            $kiekelis=0;
+            foreach ($visosp as $kk){
                 $kiekelis=$kiekelis+$kk->kiekis;
             }
             session(['kiekis'=>$kiekelis]);
@@ -80,8 +79,38 @@ class CartController extends Controller
     }
     public function deletePreke($id)
     {
+        $kr=session('krepselis');
+        if( PrekeKrepselis::where('fk_krepselis','=',$kr)->count()>1)
+        {
         PrekeKrepselis::where('id_Tarpine','=',$id)->delete();
+
+        //update krepselio suma
+            $naujakaina=0;
+            $visoskp=PrekeKrepselis::where('fk_krepselis','=',$kr)->get();
+            foreach ($visoskp as $vp){
+                $preke=Preke::where('id_preke',$vp->fk_preke)->first();
+                $naujakaina=$naujakaina+($preke->kaina*$vp->kiekis);
+            }
+            Krepselis::where('id_krepselis', $kr)->update(
+                [
+                    'kaina' => $naujakaina,
+                ]);
         return Redirect::back()->with('success', 'Item deleted');
+        }
+
+        else{
+            PrekeKrepselis::where('id_Tarpine','=',$id)->delete();
+            $visosp=PrekeKrepselis::where('fk_krepselis','=',$kr)->get();
+            Krepselis::where('id_krepselis','=',$kr)->delete();
+            $kiekelis=0;
+            foreach ($visosp as $kk){
+                $kiekelis=$kiekelis+$kk->kiekis;
+            }
+            session(['kiekis'=>$kiekelis]);
+            session()->forget('krepselis');
+
+            return Redirect::to('shop1')->with('success', 'Item deleted');
+        }
 
     }
